PUBDEV-7138 - presented model metrics can be integers

+++ h2o-algos/src/main/java/hex/tree/isoforextended/ExtendedIsolationForest.java
@@ -143,9 +143,9 @@ public class ExtendedIsolationForest extends ModelBuilder<ExtendedIsolationFores
         List<String> colTypes = new ArrayList<>();
         List<String> colFormat = new ArrayList<>();
 
-        colHeaders.add("Number of Trees"); colTypes.add("long"); colFormat.add("%d");
-        colHeaders.add("Size of Subsample"); colTypes.add("long"); colFormat.add("%d");
-        colHeaders.add("Extension Level"); colTypes.add("long"); colFormat.add("%d");
+        colHeaders.add("Number of Trees"); colTypes.add("int"); colFormat.add("%d");
+        colHeaders.add("Size of Subsample"); colTypes.add("int"); colFormat.add("%d");
+        colHeaders.add("Extension Level"); colTypes.add("int"); colFormat.add("%d");
 
         final int rows = 1;
         TwoDimTable table = new TwoDimTable(
